feat: addding session middleware

+++ models/cart.js
@@ -13,20 +13,20 @@ const p = path.join(
 
 module.exports = class Cart {
 
-  static deleteProductFromCart(id) {
-    return db.none('DELETE FROM cart WHERE product_id = $1 AND user_id = 1', id)
+  static deleteProductFromCart(product_id, user_id) {
+    return db.none('DELETE FROM cart WHERE product_id = $1 AND user_id = $2', [product_id, user_id])
   }
 
-  static getCart() {
-    return db.any('SELECT p.product_id, title, imgurl, price, description, c.qty FROM products p INNER JOIN cart c ON c.product_id = p.product_id WHERE c.user_id = 1 ORDER BY qty');
+  static getCart(user_id) {
+    return db.any('SELECT p.product_id, title, imgurl, price, description, c.qty FROM products p INNER JOIN cart c ON c.product_id = p.product_id WHERE c.user_id = $1 ORDER BY qty', user_id);
   }
 
-  static AddProductToCart(id) {
-    const promise = db.any('SELECT product_id FROM cart WHERE product_id = $1 AND user_id = 1', id).then(res => {
+  static AddProductToCart(product_id, user_id) {
+    const promise = db.any('SELECT product_id FROM cart WHERE product_id = $1 AND user_id = $2', [product_id, user_id]).then(res => {
       if (res.length > 0) {
-        return db.none('UPDATE cart SET qty = qty + 1 WHERE product_id = $1 AND user_id = 1', id)
+        return db.none('UPDATE cart SET qty = qty + 1 WHERE product_id = $1 AND user_id = $2', [product_id, user_id])
       } else {
-        return db.none('INSERT INTO cart (user_id, product_id, qty) VALUES(1, $1, 1)', id)
+        return db.none('INSERT INTO cart (user_id, product_id, qty) VALUES($2, $1, 1)', [product_id, user_id])
       }
     }).catch(err => {
       console.log(err);
@@ -34,7 +34,7 @@ module.exports = class Cart {
     return Promise.all([promise])
   }
 
-  static emptyCart() {
-    return db.none('DELETE FROM cart WHERE user_id = 1');
+  static emptyCart(user_id) {
+    return db.none('DELETE FROM cart WHERE user_id = $1', user_id);
   }
 };
